SpSeparate methods from RestAPI for the compensation module and remove unused methods
