+++ src/pages/compensation/utils/restAPI.js
@@ -1,0 +1,211 @@
+import axios, { CancelToken } from 'axios';
+
+class RestAPI {
+  
+  /** ******************* */
+  /** COMPENSATION MODULE */
+  /** ******************* */
+
+  /**
+   * Request the project impacted biomes
+   *
+   * @param {String} companyId id company to request
+   * @param {String} projectId id project to request
+   */
+  static requestImpactedBiomes(companyId, projectId) {
+    return RestAPI.makeGetRequest(`companies/${companyId}/projects/${projectId}/biomes`);
+  }
+
+  /**
+   * Request the project impacted biomes
+   *
+   * @param {String} companyId id company to request
+   * @param {String} projectId id project to request
+   */
+  static requestImpactedBiomesDecisionTree(companyId, projectId) {
+    return RestAPI.makeGetRequest(`companies/${companyId}/projects/${projectId}/decisionTree`);
+  }
+
+  /**
+   * Request available strategies information for the given parameters
+   *
+   * @param {Number} biomeId biome id
+   * @param {Number} subzoneId sub-basin id
+   * @param {String} eaId environmental authority id
+   */
+  static requestAvailableStrategies(biomeId, subzoneId, eaId) {
+    return RestAPI.makePostRequest('strategies/biomeSubzoneEA', {
+      id_biome: biomeId,
+      id_subzone: subzoneId,
+      id_ea: eaId,
+    });
+  }
+
+  /**
+   * Request the project information (props and geometry)
+   *
+   * @param {String} companyId id company to request
+   * @param {String} projectId id project to request
+   */
+  static requestProjectByIdAndCompany(companyId, projectId) {
+    return RestAPI.makeGetRequest(`companies/${companyId}/projects/${projectId}`);
+  }
+
+  /**
+   * Request the project info by company, organized by region and state
+   * @param {String} companyId id company to request
+   */
+  static requestProjectsAndRegionsByCompany(companyId) {
+    return RestAPI.makeGetRequest(`companies/${companyId}/projects?group_props=id_region,prj_status`);
+  }
+
+  /**
+   * Recover all biomes available in the database
+   */
+  static getAllBiomes() {
+    return RestAPI.makeGetRequest('biomes');
+  }
+
+  /**
+   * Create a new project
+   */
+  static createProject(companyId, regionId, statusId, name) {
+    const requestBody = {
+      name: `${name}`,
+      id_company: companyId,
+      id_region: `${regionId}`,
+      prj_status: `${statusId}`,
+      details: 'Project created by user',
+    };
+    return RestAPI.makePostRequest(`companies/${companyId}/projects`, requestBody)
+      .then((res) => ({
+        id_project: res.gid,
+        id_company: res.id_company,
+        region: res.id_region,
+        state: res.prj_status,
+        name: res.name,
+        type: 'button',
+        project: res.name.toUpperCase(),
+        label: res.name,
+        area: 0,
+      }));
+  }
+
+  /**
+   * Associate a set of biomes as impacted by the given project
+   *
+   * @param {Number} companyId company id
+   * @param {Bumber} projectId project id
+   * @param {Object[]} biomes Array of biomes info to associate
+   */
+  static addImpactedBiomesToProject(companyId, projectId, biomes) {
+    const cleanBiomes = biomes.map((biome) => ({
+      id_biome: biome.id_biome,
+      natural_area_ha: biome.natural_area_ha,
+      secondary_area_ha: biome.secondary_area_ha,
+      transformed_area_ha: biome.transformed_area_ha,
+      area_impacted_ha: biome.area_impacted_ha,
+      area_to_compensate_ha: biome.area_to_compensate_ha,
+      area_impacted_pct: biome.area_impacted_pct,
+    }));
+    return RestAPI.makePostRequest(`companies/${companyId}/projects/${projectId}/biomes`, cleanBiomes);
+  }
+
+  /**
+   * Create a new strategy as selected for the given project
+   *
+   * @param {Numer} companyId company id
+   * @param {Number} projectId project id
+   * @param {Object} strategy strategy to save information
+   */
+  static createProjectStrategy = (companyId, projectId, strategy) => RestAPI.makePostRequest(
+    `companies/${companyId}/projects/${projectId}/strategies`,
+    strategy,
+  )
+
+  /**
+   * Save many strategies as selected for the given project
+   *
+   * @param {Numer} companyId company id
+   * @param {Number} projectId project id
+   * @param {Object[]} strategies list of strategies to save
+   */
+  static bulkSaveStrategies = (companyId, projectId, strategies) => Promise.all(
+    strategies.map((strategy) => RestAPI.createProjectStrategy(companyId, projectId, strategy)),
+  )
+
+  /**
+   * Request the selected strategies for the given project
+   *
+   * @param {Numer} companyId company id
+   * @param {Number} projectId project id
+   */
+  static getSavedStrategies = (companyId, projectId) => RestAPI.makeGetRequest(
+    `companies/${companyId}/projects/${projectId}/strategies`,
+  )
+
+  /**
+   * Download the strategies saved in the given project
+   */
+  static downloadProjectStrategiesUrl = (companyId, projectId) => RestAPI.makeGetRequest(
+    `/companies/${companyId}/projects/${projectId}/strategies/download`,
+  )
+
+  /** ************** */
+  /** BASE FUNCTIONS */
+  /** ************** */
+
+  /**
+   * Request an endpoint through a GET request
+   *
+   * @param {String} endpoint endpoint to attach to url
+   */
+  static makeGetRequest(endpoint, options = {}, completeRes = false) {
+    const config = {
+      ...options,
+      headers: {
+        Authorization: `apiKey ${process.env.REACT_APP_BACKEND_KEY}`,
+      },
+    };
+    return axios.get(`${process.env.REACT_APP_BACKEND_URL}/${endpoint}`, config)
+      .then((res) => {
+        if (completeRes) {
+          return res;
+        }
+        return res.data;
+      })
+      .catch((error) => {
+        if (axios.isCancel(error)) {
+          return Promise.resolve('request canceled');
+        }
+        let message = 'Bad GET response. Try later';
+        if (error.response) message = error.response.status;
+        if (error.request && error.request.statusText === '') message = 'no-data-available';
+        return Promise.reject(message);
+      });
+  }
+
+  /**
+   * Request an endpoint through a POST request
+   *
+   * @param {String} endpoint endpoint to attach to url
+   * @param {Object} requestBody JSON object with the request body
+   */
+  static makePostRequest(endpoint, requestBody) {
+    const config = {
+      headers: {
+        Authorization: `apiKey ${process.env.REACT_APP_BACKEND_KEY}`,
+      },
+    };
+    return axios.post(`${process.env.REACT_APP_BACKEND_URL}/${endpoint}`, requestBody, config)
+      .then((res) => res.data)
+      .catch((error) => {
+        let message = 'Bad POST response. Try later';
+        if (error.response) message = error.response.status;
+        if (error.request.statusText === '') message = 'no-data-available';
+        return Promise.reject(message);
+      });
+  }
+}
+
+export default RestAPI;
