+++ src/utils/restAPI.js
@@ -17,328 +17,7 @@ class RestAPI {
     );
   }
 
-  /** ************* */
-  /** SEARCH MODULE */
-  /** ************* */
-
-  /**
-   * Recover the national area by selected strategic ecosystems
-   * @param {Number} idGeofence id geofence to request the strategic ecosystems
-   */
-  static requestNationalSE(idGeofence) {
-    return RestAPI.makeGetRequest(`se/${idGeofence}/national`);
-  }
-
-  /**
-   * Recover the national coverage by selected strategic ecosystems
-   * @param {Number} idGeofence id geofence to request the strategic ecosystems
-   */
-  static requestNationalCoverage(idGeofence) {
-    return RestAPI.makeGetRequest(`se/${idGeofence}/coverage`);
-  }
-
-  /**
-   * Request area information for biomes by subzones
-   *
-   * @param {String} eaId EA id to request
-   * @param {String} biome biome's name to request
-   */
-  static requestBiomeBySZH(eaId, biomeName) {
-    return RestAPI.makeGetRequest(`ea/${eaId}/biome/${biomeName}/subzone`);
-  }
-
-  /**
-   * Get values of functional diversity in the dry forest in a given area
-   *
-   * @param {String} areaType area type id, f.e. "ea", "states"
-   * @param {Number | String} areaId area id to request, f.e. "CRQ", 24
-   *
-   * @return {Promise<Object>} Object with values of functional diversity in the dry forest
-   */
-  static requestDryForestValues(areaType, areaId) {
-    return RestAPI.makeGetRequest(`functional-diversity/dry-forest/values?areaType=${areaType}&areaId=${areaId}`);
-  }
-
-  /**
-   * Get values of functional features in the dry forest in a given area.
-   *
-   * @param {String} areaType area type id, f.e. "ea", "states"
-   * @param {Number | String} areaId area id to request, f.e. "CRQ", 24
-   *
-   * @return {Promise<Array>} Array of objects with values of functional features in the dry forest
-   */
-  static requestDryForestFeatures(areaType, areaId) {
-    return RestAPI.makeGetRequest(`functional-diversity/dry-forest/features?areaType=${areaType}&areaId=${areaId}`);
-  }
-
-  /**
-   * Get texts associated to one section
-   *
-   * @param {String} key section key
-   *
-   * @return {Promise<Object>} Object with texts
-   */
-   static requestSectionTexts(key) {
-    return RestAPI.makeGetRequest(`util/texts?key=${key}`);
-  }
-
-  /** ******************** */
-  /** MAPS - SEARCH MODULE */
-  /** ******************** */
-
-  /**
-   * Request area geometry by id
-   *
-   * @param {String} areaId area id to request
-   *
-   * @return {Object} Including Promise with layer object to load in map and source reference to
-   * cancel the request
-   */
-  static requestNationalGeometryByArea(areaId) {
-    const source = CancelToken.source();
-    return {
-      request: RestAPI.makeGetRequest(`${areaId}/layers/national`, { cancelToken: source.token }),
-      source,
-    };
-  }
-
-  /**
-   * Get the layer of number of species for the specified group
-   *
-   * @param {String} areaType area type id, f.e. "ea", "states"
-   * @param {Number | String} areaId area id to request, f.e. "CRQ", 24
-   * @param {String} group group to get the layer for, options are: total | endemic | invasive |
-   * threatened
-   *
-   * @return {Promise<Object>} layer object to be loaded in the map
-   */
-  static requestNOSLayer(areaType, areaId, group) {
-    const source = CancelToken.source();
-    return {
-      request: RestAPI.makeGetRequest(
-        `richness/number-species/layer?areaType=${areaType}&areaId=${areaId}&group=${group}`,
-        { cancelToken: source.token, responseType: 'arraybuffer' },
-        true,
-      ),
-      source,
-    };
-  }
-
-  /**
-   * Get the threshold values for the layer of number of species for the specified group
-   *
-   * @param {String} areaType area type id, f.e. "ea", "states"
-   * @param {Number | String} areaId area id to request, f.e. "CRQ", 24
-   * @param {String} group group to get the layer for, options are: total | endemic | invasive |
-   * threatened
-   *
-   * @return {Promise<Object>} object with min an max values
-   */
-  static requestNOSLayerThresholds(areaType, areaId, group) {
-    return RestAPI.makeGetRequest(
-      `richness/number-species/layer/thresholds?areaType=${areaType}&areaId=${areaId}&group=${group}`,
-    );
-  }
-
-  /**
-   * Request data available to custom geometry
-   *
-   * @param {Object} polygon polygon saved
-   *
-   * @return {Object} with data related to the polygon
-   */
-  static requestCustomPolygonData(polygon) {
-    /** TODO: implement all this endpoint and also the backend response
-     * to find information according to polygon coordinates
-     * */
-    return RestAPI.makePostRequest('polygon', {
-      latLngs: polygon.latLngs,
-    });
-  }
-
-  /**
-   * Get the layer of gaps section of richness
-   *
-   * @param {String} areaType area type id, f.e. "ea", "states"
-   * @param {Number | String} areaId area id to request, f.e. "CRQ", 24
-   *
-   * @return {Promise<Object>} layer object to be loaded in the map
-   */
-  static requestGapsLayer(areaType, areaId) {
-    const source = CancelToken.source();
-    return {
-      request: RestAPI.makeGetRequest(
-        `richness/gaps/layer?areaType=${areaType}&areaId=${areaId}`,
-        { cancelToken: source.token, responseType: 'arraybuffer' },
-        true,
-      ),
-      source,
-    };
-  }
-
-  /**
-   * Get the threshold values for the layer of gaps section of richness
-   *
-   * @param {String} areaType area type id, f.e. "ea", "states"
-   * @param {Number | String} areaId area id to request, f.e. "CRQ", 24
-   *
-   * @return {Promise<Object>} object with min an max values
-   */
-  static requestGapsLayerThresholds(areaType, areaId) {
-    return RestAPI.makeGetRequest(
-      `richness/gaps/layer/thresholds?areaType=${areaType}&areaId=${areaId}`,
-    );
-  }
-
-  /** ******************* */
-  /** COMPENSATION MODULE */
-  /** ******************* */
-
-  /**
-   * Request the project impacted biomes
-   *
-   * @param {String} companyId id company to request
-   * @param {String} projectId id project to request
-   */
-  static requestImpactedBiomes(companyId, projectId) {
-    return RestAPI.makeGetRequest(`companies/${companyId}/projects/${projectId}/biomes`);
-  }
-
-  /**
-   * Request the project impacted biomes
-   *
-   * @param {String} companyId id company to request
-   * @param {String} projectId id project to request
-   */
-  static requestImpactedBiomesDecisionTree(companyId, projectId) {
-    return RestAPI.makeGetRequest(`companies/${companyId}/projects/${projectId}/decisionTree`);
-  }
-
-  /**
-   * Request available strategies information for the given parameters
-   *
-   * @param {Number} biomeId biome id
-   * @param {Number} subzoneId sub-basin id
-   * @param {String} eaId environmental authority id
-   */
-  static requestAvailableStrategies(biomeId, subzoneId, eaId) {
-    return RestAPI.makePostRequest('strategies/biomeSubzoneEA', {
-      id_biome: biomeId,
-      id_subzone: subzoneId,
-      id_ea: eaId,
-    });
-  }
-
-  /**
-   * Request the project information (props and geometry)
-   *
-   * @param {String} companyId id company to request
-   * @param {String} projectId id project to request
-   */
-  static requestProjectByIdAndCompany(companyId, projectId) {
-    return RestAPI.makeGetRequest(`companies/${companyId}/projects/${projectId}`);
-  }
-
-  /**
-   * Request the project info by company, organized by region and state
-   * @param {String} companyId id company to request
-   */
-  static requestProjectsAndRegionsByCompany(companyId) {
-    return RestAPI.makeGetRequest(`companies/${companyId}/projects?group_props=id_region,prj_status`);
-  }
-
-  /**
-   * Recover all biomes available in the database
-   */
-  static getAllBiomes() {
-    return RestAPI.makeGetRequest('biomes');
-  }
-
-  /**
-   * Create a new project
-   */
-  static createProject(companyId, regionId, statusId, name) {
-    const requestBody = {
-      name: `${name}`,
-      id_company: companyId,
-      id_region: `${regionId}`,
-      prj_status: `${statusId}`,
-      details: 'Project created by user',
-    };
-    return RestAPI.makePostRequest(`companies/${companyId}/projects`, requestBody)
-      .then((res) => ({
-        id_project: res.gid,
-        id_company: res.id_company,
-        region: res.id_region,
-        state: res.prj_status,
-        name: res.name,
-        type: 'button',
-        project: res.name.toUpperCase(),
-        label: res.name,
-        area: 0,
-      }));
-  }
-
-  /**
-   * Associate a set of biomes as impacted by the given project
-   *
-   * @param {Number} companyId company id
-   * @param {Bumber} projectId project id
-   * @param {Object[]} biomes Array of biomes info to associate
-   */
-  static addImpactedBiomesToProject(companyId, projectId, biomes) {
-    const cleanBiomes = biomes.map((biome) => ({
-      id_biome: biome.id_biome,
-      natural_area_ha: biome.natural_area_ha,
-      secondary_area_ha: biome.secondary_area_ha,
-      transformed_area_ha: biome.transformed_area_ha,
-      area_impacted_ha: biome.area_impacted_ha,
-      area_to_compensate_ha: biome.area_to_compensate_ha,
-      area_impacted_pct: biome.area_impacted_pct,
-    }));
-    return RestAPI.makePostRequest(`companies/${companyId}/projects/${projectId}/biomes`, cleanBiomes);
-  }
-
-  /**
-   * Create a new strategy as selected for the given project
-   *
-   * @param {Numer} companyId company id
-   * @param {Number} projectId project id
-   * @param {Object} strategy strategy to save information
-   */
-  static createProjectStrategy = (companyId, projectId, strategy) => RestAPI.makePostRequest(
-    `companies/${companyId}/projects/${projectId}/strategies`,
-    strategy,
-  )
-
-  /**
-   * Save many strategies as selected for the given project
-   *
-   * @param {Numer} companyId company id
-   * @param {Number} projectId project id
-   * @param {Object[]} strategies list of strategies to save
-   */
-  static bulkSaveStrategies = (companyId, projectId, strategies) => Promise.all(
-    strategies.map((strategy) => RestAPI.createProjectStrategy(companyId, projectId, strategy)),
-  )
-
-  /**
-   * Request the selected strategies for the given project
-   *
-   * @param {Numer} companyId company id
-   * @param {Number} projectId project id
-   */
-  static getSavedStrategies = (companyId, projectId) => RestAPI.makeGetRequest(
-    `companies/${companyId}/projects/${projectId}/strategies`,
-  )
-
-  /**
-   * Download the strategies saved in the given project
-   */
-  static downloadProjectStrategiesUrl = (companyId, projectId) => RestAPI.makeGetRequest(
-    `/companies/${companyId}/projects/${projectId}/strategies/download`,
-  )
-
+  
   /** ************** */
   /** BASE FUNCTIONS */
   /** ************** */
